For the sake of simplicity it is better to get rid of `AttachmentsReadbackGroup`, and instead add multiple attachment read back feature to the `AttahcmentReadback` class.
